backendPatitasChile/requirements.txt
Django==4.2.3
djangorestframework==3.15.2
django-cors-headers==4.6.0
drf-yasg==1.21.8
djangorestframework-simplejwt==5.3.1
pillow == pillow 11.0.0